requirements.txt
# ┌─────────────────────────────┐
# │ compare_ocr 固定依赖        │
# │ Python 3.10 / Win & mac     │
# └─────────────────────────────┘
paddlepaddle==2.6.3
numpy==1.24.4
opencv-python-headless==4.6.0.66
PyMuPDF==1.22.3
pillow>=10.2,<11.0
pdf2docx==0.5.8
pandas==2.2.2
rapidfuzz==3.13.0
scipy==1.15.3
# 打包机需用 PyInstaller；开发机可选装
pyinstaller==6.6.0
